fix remaining tests
