+++ envergo/moulinette/tests/test_optional_criteria.py
@@ -3,6 +3,7 @@ from django.urls import reverse
 from pytest_django.asserts import assertTemplateUsed
 
 from envergo.geodata.conftest import france_map  # noqa
+from envergo.geodata.conftest import mock_geo_api_data  # noqa
 from envergo.moulinette.tests.factories import (
     CriterionFactory,
     MoulinetteConfigFactory,
@@ -46,7 +47,9 @@ def moulinette_data(footprint):
 
 
 # ETQ Admin, je peux voir l'option d'activer un critère optionnel
-def test_admin_see_optional_criterion_additional_question(admin_client):
+def test_admin_see_optional_criterion_additional_question(
+    admin_client, mock_geo_api_data  # noqa
+):
     url = reverse("moulinette_result")
     params = "created_surface=500&final_surface=500&lng=-1.54394&lat=47.21381"
     full_url = f"{url}?{params}"
@@ -66,7 +69,9 @@ def test_admin_see_optional_criterion_additional_question(admin_client):
 
 
 # ETQ User, je ne peux pas voir l'option d'activer un critère optionnel
-def test_users_cannot_see_optional_criterion_additional_question(client):
+def test_users_cannot_see_optional_criterion_additional_question(
+    client, mock_geo_api_data  # noqa
+):
     url = reverse("moulinette_result")
     params = "created_surface=500&final_surface=500&lng=-1.54394&lat=47.21381"
     full_url = f"{url}?{params}"
@@ -83,7 +88,7 @@ def test_users_cannot_see_optional_criterion_additional_question(client):
 
 
 # ETQ Admin, je peux consulter une simulation avec un critère optionnel
-def test_admin_see_optional_criterion_result(admin_client):
+def test_admin_see_optional_criterion_result(admin_client, mock_geo_api_data):  # noqa
     url = reverse("moulinette_result")
     params = "created_surface=500&final_surface=500&lng=-1.54394&lat=47.21381&evalenv_rubrique_41-activate=on&evalenv_rubrique_41-nb_emplacements=gte_50&evalenv_rubrique_41-type_stationnement=public"  # noqa
     full_url = f"{url}?{params}"
@@ -100,7 +105,7 @@ def test_admin_see_optional_criterion_result(admin_client):
 
 
 # ETQ User, je ne peux pas consulter une simulation avec un critère optionnel
-def test_users_cannot_see_optional_criterion_results(client):
+def test_users_cannot_see_optional_criterion_results(client, mock_geo_api_data):  # noqa
     url = reverse("moulinette_result")
     params = "created_surface=500&final_surface=500&lng=-1.54394&lat=47.21381&evalenv_rubrique_41-activate=on&evalenv_rubrique_41-nb_emplacements=gte_50"  # noqa
     full_url = f"{url}?{params}"
@@ -110,7 +115,7 @@ def test_users_cannot_see_optional_criterion_results(client):
     assert "evalenv_rubrique_41_soumis" not in res["Location"]
 
 
-def test_optional_criterion_activation(admin_client):
+def test_optional_criterion_activation(admin_client, mock_geo_api_data):  # noqa
     url = reverse("moulinette_result")
     params = "created_surface=500&final_surface=500&lng=-1.54394&lat=47.21381&evalenv_rubrique_41-activate=on"
     full_url = f"{url}?{params}"
