+++ envergo/moulinette/tests/test_views.py
@@ -1,9 +1,8 @@
-from unittest.mock import Mock, patch
-
 import pytest
 from django.urls import reverse
 from pytest_django.asserts import assertTemplateUsed
 
+from envergo.geodata.conftest import mock_geo_api_data  # noqa
 from envergo.moulinette.tests.factories import MoulinetteConfigFactory
 
 pytestmark = pytest.mark.django_db
@@ -18,47 +17,9 @@ UNAVAIL = "Le simulateur EnvErgo n'est pas encore déployé dans votre départem
 ADMIN_MSG = "Le simulateur n'est pas activé dans ce département"
 
 
-@pytest.fixture(autouse=True)
-def mock_geo_api_data():
-    with patch(
-        "envergo.geodata.utils.get_data_from_coords", new=Mock()
-    ) as mock_geo_data:
-        mock_geo_data.return_value = {
-            "type": "housenumber",
-            "name": "10 La Pommeraie",
-            "label": "10 La Pommeraie 44140 Montbert",
-            "street": "La Pommeraie",
-            "postcode": "44140",
-            "citycode": "44102",
-            "city": "Montbert",
-            "oldcitycode": None,
-            "oldcity": None,
-            "context": "44, Loire-Atlantique, Pays de la Loire",
-            "importance": 0.47452,
-            "housenumber": "10",
-            "id": "44102_haa6rn_00010",
-            "x": 359347.63,
-            "y": 6670527.5,
-            "distance": 78,
-            "score": 0.9922,
-            "_type": "address",
-        }
-        yield mock_geo_data
-
-
-@pytest.fixture(autouse=True)
-def mock_geo_api_commune():
-    with patch(
-        "envergo.geodata.utils.get_commune_from_coords", new=Mock()
-    ) as mock_commune:
-        mock_commune.return_value = {"code": "44102", "nom": "Montbert"}
-        yield mock_commune
-
-
 def test_moulinette_home(client):
     url = reverse("moulinette_home")
     res = client.get(url)
-
     assert res.status_code == 200
     assert HOME_TITLE in res.content.decode()
     assert RESULT_TITLE not in res.content.decode()
@@ -74,7 +35,7 @@ def test_moulinette_home_with_params_redirects_to_results_page(client):
     assert res.url.startswith("/simulateur/resultat/")
 
 
-def test_moulinette_result_without_config(client):
+def test_moulinette_result_without_config(client, mock_geo_api_data):  # noqa
     """When dept. contact info is not set, eval is unavailable."""
 
     url = reverse("moulinette_result")
@@ -86,7 +47,9 @@ def test_moulinette_result_without_config(client):
     assertTemplateUsed(res, "moulinette/result_non_disponible.html")
 
 
-def test_moulinette_result_without_config_admin_access(client, admin_user):
+def test_moulinette_result_without_config_admin_access(
+    client, admin_user, mock_geo_api_data  # noqa
+):
     """When dept. contact info is not set, eval is unavailable, even for admins."""
     client.force_login(admin_user)
 
@@ -99,7 +62,7 @@ def test_moulinette_result_without_config_admin_access(client, admin_user):
     assertTemplateUsed(res, "moulinette/result_non_disponible.html")
 
 
-def test_moulinette_result_with_deactivated_config(client):
+def test_moulinette_result_with_deactivated_config(client, mock_geo_api_data):  # noqa
     MoulinetteConfigFactory(is_activated=False)
 
     url = reverse("moulinette_result")
@@ -111,7 +74,9 @@ def test_moulinette_result_with_deactivated_config(client):
     assertTemplateUsed(res, "moulinette/result_available_soon.html")
 
 
-def test_moulinette_result_with_deactivated_config_admin_access(client, admin_user):
+def test_moulinette_result_with_deactivated_config_admin_access(
+    client, admin_user, mock_geo_api_data  # noqa
+):
     MoulinetteConfigFactory(is_activated=False)
     client.force_login(admin_user)
 
@@ -125,7 +90,7 @@ def test_moulinette_result_with_deactivated_config_admin_access(client, admin_us
     assert ADMIN_MSG in res.content.decode()
 
 
-def test_moulinette_result_with_activated_config(client):
+def test_moulinette_result_with_activated_config(client, mock_geo_api_data):  # noqa
     MoulinetteConfigFactory(is_activated=True)
 
     url = reverse("moulinette_result")
